+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Changes:
 
+* feature: validate certificates that users want to use as a `provided` CA
+  [#565](https://github.com/Kong/kuma/pull/565)
 * fix: add MADS port to K8S install script
   [#564](https://github.com/Kong/kuma/pull/564)
 * feature: sanitize metrics for StatsD and Prometheus